refactor: rename files to `editor.*` and remove old `types.ts`
