+++ client-v2/src/app/components/organisms/task/task.component.ts
@@ -35,8 +35,8 @@ import {
     getDefaultEditorFeatures,
     getDefaultEditorLayout,
     provideEditorFeatures,
-} from 'src/app/rich-text-editor/features'
-import { ChecklistCount, countChecklistItems } from 'src/app/rich-text-editor/helpers'
+} from 'src/app/rich-text-editor/editor.features'
+import { ChecklistCount, countChecklistItems } from 'src/app/rich-text-editor/editor.helpers'
 import { TipTapEditorComponent } from 'src/app/rich-text-editor/tip-tap-editor/tip-tap-editor.component'
 import { DeviceService } from 'src/app/services/device.service'
 import { colors, taskStatusColorMap } from 'src/app/shared/colors'
