+++ client-v2/src/app/rich-text-editor/editor-features/search-and-replace.feature.ts
@@ -1,6 +1,6 @@
 import { SearchAndReplace } from './extensions/search-and-replace.extension'
 import { EditorFeatureId } from '../editor.types'
-import { createEditorFeature } from '../helpers'
+import { createEditorFeature } from '../editor.helpers'
 
 export const searchAndReplaceFeature = createEditorFeature({
     featureId: EditorFeatureId.SearchAndReplace,
