+++ client-v2/src/app/rich-text-editor/editor.helpers.ts
